// src/types/Word.ts
export class Word {
  id: Number;
  syn: Number;
  word: String;
  roots: String;
  r1: Number;
  d1: Number;
  r2: Number;
  d2: Number;
  r3: Number;
  d3: Number;
  r4: Number;
  d4: Number;
  r5: Number;
  d5: Number;
  mix: Number;
  param: Number;
  constructor () {
    this.id = 0;
    this.syn = 0;
    this.word = "";
    this.roots = "";
    this.r1 = 0;
    this.d1 = 0;
    this.r2 = 0;
    this.d2 = 0;
    this.r3 = 0;
    this.d3 = 0;
    this.r4 = 0;
    this.d4 = 0;
    this.r5 = 0;
    this.d5 = 0;
    this.mix = 1;
    this.param = 0;
  }
}